fix(audit): pass 11 — show leg reverses time, not textures
MEDIUM: the show-back leg swapped the two captures and drove progress
forward, which only equals a reverse for a SYMMETRIC pack. Peek Recede's
return leg therefore shrank the WALLPAPER away to reveal static windows
instead of growing the windows back, and Phosphor Peek drained the
desktop rather than re-energizing the circuit — while the CHANGELOG
promised 'plays it in reverse' and all three pack headers claimed FROM
is always the windows scene. Both legs now share one endpoint pair
(FROM = windows scene, TO = bare desktop) and differ only in direction:
hide drives t 0 to 1, show drives it 1 to 0. The show leg captures its
windows endpoint live (KWin unhides before the signal) and takes its
desktop endpoint from the hide leg's cache, so an asymmetric pack
retraces its own motion — matching windowaperture's redirect-Backward
return leg.

MEDIUM: the ToVirtual free-geometry remap was the last loop without the
already-valid-VS guard its four siblings carry (pass 10 widened that gap
from 3-of-4 to 4-of-5); a VS reconfiguration could relocate a still-valid
vs:1 float-back rect to vs:0 and force a no-op save.

LOW: desktoptransitionmanager.cpp had grown back to 813 lines, so its
teardown half (settle, reap, removals, release) moves to
desktoptransitionteardown.cpp — the drive/capture/assembly/teardown split
now leaves every TU at 645/353/196/121; stateForWindow's docstring said
nullptr-when-untracked when it never returns null once wired (the origin
of the dead null-checks); the animation min-size loaders' comment claimed
a symmetry with animationDuration that pass 10 inverted.

NIT: an unreachable Q_ASSERT is gone (insertTransition asserts the same
contract itself); its reasoning moves into the rollback comment, which
is where the no-safe-action trade actually needs recording.
