+++ libs/phosphor-snap-engine/include/PhosphorSnapEngine/SnapEngine.h
@@ -432,8 +432,13 @@ public:
     // as a per-window value, updated in place, exactly as the former single store.
     // ═══════════════════════════════════════════════════════════════════════════
 
-    /// The SnapState that owns @p windowId (via the reverse map), or nullptr when
-    /// the window is untracked. Never creates.
+    /// The SnapState that owns @p windowId (via the reverse map), or the global
+    /// holder when the window is untracked — NEVER null once the engine is
+    /// wired (see the fallback rationale on the definition). An untracked
+    /// window then reads empty per-window data out of the holder, which is
+    /// observably identical to the nullptr this once returned. Never creates.
+    /// Callers must not null-check the result as a tracked/untracked test; ask
+    /// the reverse map (or compare against globalState()) instead.
     SnapState* stateForWindow(const QString& windowId);
     const SnapState* stateForWindow(const QString& windowId) const;
